Added lots of commenting. Still lots to go in DBAccess.
Added the ability to get the current user and character. Seems to be
working at least for CommandSample. I haven't tested it with throwaway
accounts yet though.

Read the javadocs on the methods in DBAccess! If it has well formatted
javadoc, it is probably meant to be used by the ODP. Some methods will
still be moved out however, while other methods will be added in.