Little grammatical fix
Before this commit, if an attack was blocked by 2 pieces of armor, it would print as    
`The attack was {partially/completely} blocked due to the Armor1, and Armor2` - that is grammatically wrong and always irked me a bit whenever I saw it.   

Also why don't you use StringBuilder or StringBuffer for this?