Changes to pickup actions
Added some features to ODPDBAccess doMoveItem() and tryMoveItem(). Before, when the total weight of a stack of items was too heavy to carry, you couldn't pick up any of the items at all. Now, the stack is split and whatever amount you can carry is added to your inventory. Transaction wrapper added to prevent item duplication.