Renamed the old "Exit" button that would show up on combat sites to "Leave and forget".
For locations with only a single exit, you now see that exit as a link on your banner.

Fixed a bug in the experimental UI on mobile causing the whole screen to go a little cra-cra.

Updated the Experimental UI a bit to make it more usable (especially on desktop). You'll find the popup windows now stay on the right half of the screen and don't require scrolling down past the chat. The game window scrunches 50% to fit it in, maximizing realestate efficiency imo. Mobile is unchanged (in theory). Also, popup messages no longer just blur the screen and screw you up permanently. It is very likely that the combat experience in this UI mode is not yet usable but I haven't tested it yet.